Allow for building prereleases
"release" gradle project property may be set to "true" when building mod, this causes the mod to build as a release version; prerelease build is the default option
The prerelease build causes the .jar file to have the commit hash in it's name instead of the version.
The hash will also be displayed in the mod version in /twitch status.
In prerelease builds, the update checker will look not only for different version in gradle.properties, but also for a new commit hash on the master branch