Convert current implementation of message redirection to FML event
Instead of GuiScreenHook.redirectMessage directly managing the redirection, it now fires a new event (LocalMessageEvent)
This event can be cancelled and the message content may be edited.
Cancelling the event blocks the message, editing the message field edits the final message.
The implementation of message redirection is now in StreamEvents.
If the Twitch chat is disabled or no channel is selected while redirection is enabled, the message is now still blocked, but an error is shown.