Some emote registering optimizations & error propagation
Animated emotes with large amounts of frames should load faster at runtime.
Previously, frames would be registered one at a time.
This isn't an issue when in the mod loading phase, but it becomes one when trying to load emotes at runtime.
This is because each texture must be registered in the main thread, but runtime emote registration is done in a separate thread to avoid blocking the main game loop.
In order to register a texture, we must create a task on the main thread, then wait for it to finish.
Pending tasks on the main thread are processed once every tick (20 times a second).
So in the old system of registering frames one by one, an emote with 200 frames would take **10 seconds** to register.
The new method first generates all textures, then registers them all in one tick.